[finishes #162375145] refactor and create middlewares
