+++ package.json
@@ -27,23 +27,24 @@
   },
   "homepage": "https://github.com/vic3king/iReporter#readme",
   "dependencies": {
-   "babel-cli": "^6.26.0",
+    "babel-cli": "^6.26.0",
     "babel-preset-env": "^1.7.0",
     "babel-watch": "^2.0.7",
+    "bcryptjs": "^2.4.3",
     "chai": "^4.2.0",
     "chai-http": "^4.2.0",
     "coveralls": "^3.0.2",
     "eslint": "^5.9.0",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.14.0",
+    "express": "^4.16.4",
     "istanbul": "^0.4.5",
+    "jsonwebtoken": "^8.4.0",
     "mocha": "^5.2.0",
     "nyc": "^13.1.0",
     "travis": "^0.1.1",
-    "express": "^4.16.4",
-    "uuid": "^3.3.2"
+    "uuid": "^3.3.2",
+    "validator": "^10.9.0"
   },
-  "devDependencies": {
-   
-  }
+  "devDependencies": {}
 }
